Replace attributes map by AttributeKeyValue pair list
This is the first of commits that implements RFC0059 [1]. I will submit changes introduced
in the RFC in separate PRs for easier review.

This change is already benchmarked as part of the RFC, so I will not submit individual benchmarks
with each PR.

Note: this change has been thoroughly reviewed, commented and discussed as part of the RFC
process. This commit merely implements part of the RFC.

[1] - https://github.com/open-telemetry/oteps/blob/master/text/0059-otlp-trace-data-format.md