Refine child span count field
The field was previously defined ambiguously and it was not clear
whether it was the count of local or all children (including remotely
generated).

This change makes it clear that this is only local children. A counter
of remote children may be added in the future if there is an understanding
about how to obtain this number (open discussion is here:
https://github.com/open-telemetry/opentelemetry-specification/issues/355).

The field type is also changed to a more efficient sfixed32 with
special-casing -1 as "value not known".