Eliminate multi-data point timeseries
Resolves https://github.com/open-telemetry/opentelemetry-proto/issues/54

After a discussion we came to conclusion that there is no good use case
for having multiple data points. All possible ways to combine data points
that we imagined are either very expensive or make no semantic sense.

This commit eliminates timeseries that can have multiple data points and replaces
them with single data points.

If in the future we discover a use case for multiple data points we have 2
possible paths to extend the proto:

- Add a corresponding points repeated field in the Int64DataPoint.
- Introduce a different message type that is capable of storing multiple data
  points and assign it a different type in the MetricDescriptor.Type enum.

There are probably other ways as well.