Declare Trace, Resource and Common as Beta
This adds the maturity levels for individual parts of the protocol
according to recently agreed project-wide definition of maturity levels.

After merging this change we will declare Trace (Span), Resource and
Common parts of the protocol as Beta, which means no more than once in
3 months changes are allowed after that.

We need this declaration of maturity so that we can begin implementation
of OTLP in the Collector.

One known pending change that prevents us from declaring the components
as Stable is https://github.com/open-telemetry/opentelemetry-proto/pull/94

We assume that there will be no other changes within the next 3-months timeframe.
The assumption seems reasonable since there are no other known shortcomings
in this part of the protocol that may require more modifications.

Note that metrics part is considered Alpha since it is still under active
development.