Rename Span Event name to short_name
This is part of the alignment effort defined here https://github.com/open-telemetry/opentelemetry-specification/issues/622

The comment added clarifies the semantics, but I believe this is compatible
with our previous understanding and doe not change that understanding.

This does not break the protocol since renaming the fields does not change
the wire format.