Add Monotonic Types (#145)
* Add Monotonic refinement

Knowing the monotonicity of metric data enables receivers to implement
certain algorithms to calculate the rate of change of the data. As such,
this is an important quality to describe and is added to the
`MetricDescriptor`.

*Note on English*: NONDECREASING is used to describe values that either
increase or remain constant over time. The English term "increasing" is
not used as it does not encompass values that remain constant over time.

A comment is added to identify the omission of NONINCREASING and
information on how to notify it is needed should that arise in the
future.

* Monotonic in type

Instead of including a monotonic refinement, specify where applicable to
the type.

Co-authored-by: Bogdan Drutu <bogdandrutu@gmail.com>