Remove Temporality and add Structure
Consensus was reached that the description of the time interval
over which data was measured can be determined as `DELTA` or
`CUMULATIVE` by comparison of the timestamps sent with the data. The
classification of `INSTANTANEOUS` has also been found irrelevant in
meaningfully description of the data. For these reasons, `Temporality` is
removed.

Conversely, an understanding of what operations can be performed on data is
essential information to contain in this protocol so it can support systems
of statistics or roll-up. To provide this support, the `Structure` enum is
added and an appropriate field in the Metric descriptor included to capture
if the data is adding or grouping.