Declare common and resource directories stable
We declared traces part of the protocol stable. Traces use common
and resource messages so we have to declare these as stable as well
otherwise the declaration of trace part as stable does not achieve
the goal (the stable interoperability).

Obviously any new messages can be freely added to common and resource,
"stable" label does not prevent that.