Deprecate old explicit bounds, add message for explicit bounds
This is required in order for the protocol to develope in the future and support adding
different other types of bounds. With this change it is backwards compatible to add multiple
bound types and embed them into an oneof (wire compatible).

On any service (receiver) if the deprecated field is received the only operation required
is to reassign the deprecated field to the new embeded repeated field in the message.

Signed-off-by: Bogdan Drutu <bogdandrutu@gmail.com>