Change OTLP/HTTP port from 4317 to 4318 (#1839)
Related to https://github.com/open-telemetry/opentelemetry-specification/issues/1816
Fixes https://github.com/open-telemetry/opentelemetry-specification/issues/1835

Some historical context: we wanted to make grpc and http use the same port and we
had an open issue in the Collector to do so:
https://github.com/open-telemetry/opentelemetry-collector/issues/1256

The conclusion is that there are technical hurdles that make it unfeasible:
https://github.com/open-telemetry/opentelemetry-collector/issues/1256#issuecomment-880048833

Because of that we need to keep grpc and http ports separate. This means we need to
change the spec to say that otlp/http uses port 4318. Once this PR is merged we
will also need to submit for port 4318 registration with IANA like we did previously
with port 4317:
https://github.com/open-telemetry/opentelemetry-specification/issues/1148#issuecomment-725067404