Rename InstrumentationLibrary to InstrumentationScope
This applies changes described in spec PR https://github.com/open-telemetry/opentelemetry-specification/pull/2276

This is not a breaking change for OTLP. The wire format does not change.
It is a breaking change of the names of the generated code, which is acceptable
since we are still on 0.x version of this repository.