Rename InstrumentationLibrary to InstrumentationScope (#362)
This applies changes described in spec PR https://github.com/open-telemetry/opentelemetry-specification/pull/2276

This is not a breaking change for OTLP. The wire format does not change.

The change is done in a way that the transition can be handled gracefully for OTLP/JSON and for the generated code. See the comments for explanation on how to do it in the senders and receivers.