Delete unused experimental TraceConfig
Resolves https://github.com/open-telemetry/opentelemetry-proto/issues/410

I am not aware of any OpenTelemetry or non-OpenTelemetry
codebase that uses the experimental TraceConfig.

I suggest to remove it. It can be added back if we see the
need for it in the future.

The SDK configuration may also be achieved by different means,
see https://github.com/open-telemetry/opentelemetry-specification/issues/2207

We need to keep this PR open for a while and circulate it widely
to make sure we give enough opportunity to object the removal.