Remove helper masks, not part of the protocol
Fixes https://github.com/open-telemetry/opentelemetry-proto/issues/363

The reason is to keep the proto interface as minimal as possible, and the helper masks are just helpers, and not part of the protocol. They can be defined by the language SIGs in their usages, but don't need to be part of the proto stability.

Signed-off-by: Bogdan Drutu <bogdandrutu@gmail.com>