Clarify how additive changes are handled
This PR explicitly lists the additive changes allowed and adds
a requirement that such additive changes must be accompanied by
interoperability explanation when necessary.

This is a subset of https://github.com/open-telemetry/opentelemetry-proto/pull/432
that contains other guarantees that we did not yet agree to.
I believe this particular subset is necessary regardless of
what we decide about #432.