Clarify that symbolic guarantees go into effect from 1.0.0
We merged the https://github.com/open-telemetry/opentelemetry-proto/pull/432
without clarifying when exactly the guarantees start applying. The intent
was that the guarantees are needed for 1.0 release but we didn't make this
explicit.

We are now making this explicit and until 1.0 released we should be free
to fix the remaining known issues listed here:
https://github.com/open-telemetry/opentelemetry-proto/issues/456

Once 1.0 release is made we can remove the [from 1.0.0] labels from the README.md.