Clarify experimental additions to Stable components
The Entities SIG is planning a backward-compatible addition to the Resource,
see the prototype here: https://github.com/tigrannajaryan/opentelemetry-proto/pull/2

The change is compliant with existing stability requirements in this repository.

However, our stability definition does not specify how such experimental additions
to stable signals are expected to be done and what happens during and after
the experiment.

This PR adds more clarity in the process and further refines the stability
definition to reinforce our strong commitment ot OTLP's stability.