Clarify when bucket_counts array may be empty
The current definition says:

"The number of elements in bucket_counts array must be by one
 greater than the number of elements in explicit_bounds array."

However, it also says "bucket_counts is an optional field",
which means its length can be 0, which than in turn means the
length of explicit_bounds must be -1, which is impossible.

We clarify that it is a valid situation to have both bucket_counts
and explicit_bounds arrays empty.