Optimized profile stack trace representation.
In #645 a simplified stack trace representation is proposed. A
discussion there asks for a more efficient way to represent stacks that
slightly differ at the leaf. A specialized solution proposed there would
put the leaf location as a separate field. That seems partial and
potentially error-prone.

This PR proposes an alternative stack trace representation, shaped as
a tree in the form of two arrays. See the code for details and an
example.

The advantage of this approach is that stacks with the same prefix are
encoded more compactly. Also, since we use only two arrays for the
encoding, the in-memory representation is also more efficient in terms
of the number of allocations required. The main disadvantage is that
this approach is more complex semantically and for this reason can be
more error-prone.