profiles: improve sample_type comment
I think using nanoseconds is the better default unit for CPU profiles
because the weight of a count can be different depending on the sample
rate.

Using the same unit for off_cpu and cpu is more consistent.

The comment above was a little confusing, since it referred only to a
cpu profile, but was then followed by an off_cpu profile sample as well.