Document more explicitly the dictionary zero element requirement.
Right now we have a comment to the dictionary proto and a comment to the
string_table field. This requirement is important enough I think that
it's worth documenting it more explicitly, especially the part that even
if there are no references to the particular dictionary, the zero index
element must still be present.