Rename line -> lines and sample -> samples since these are repeated fields.
Protobuf
[docs](https://protobuf.dev/programming-guides/style/#field-names) say:

> Use pluralized names for repeated fields.

Other fields in profiles.proto and other OTLP types follow that
recommended style AFAICT. The only fields that do not are `line` and
`sample`, and I think those are remnants of the now-gone exact pprof
format compatibility we explored early on.

As we approach the Alpha release of the profiles signal, let's be
consistent in how we name repeated fields and use plural here.